refactor: coalesce zod utils into "parse" frame
