+++ packages/common-all/src/parse.ts
@@ -1,7 +1,21 @@
-import { z } from "./util/zodUtil";
+import { z } from "zod";
+import type { ZodType } from "zod";
 import { ok, err } from "./utils";
 import type { Result } from "neverthrow";
 import { DendronError } from "./error";
+
+export { z };
+
+/**
+ * util for defining zod schemas with external/custom types.
+ * Origin: https://github.com/colinhacks/zod/issues/372#issuecomment-826380330
+ * @returns a function to be called with a zod schema
+ */
+export const schemaForType =
+  <T>() =>
+  <S extends ZodType<T, any, any>>(arg: S) => {
+    return arg;
+  };
 
 /**
  * Parse `zod` schema into `Result`
