+++ packages/common-all/src/types/configs/publishing/publishing.ts
@@ -2,7 +2,7 @@ import { DVault } from "../../DVault";
 import { GiscusConfig } from "./giscus";
 import { GithubConfig, genDefaultGithubConfig, githubSchema } from "./github";
 import { SEOConfig, genDefaultSEOConfig, seoSchema } from "./seo";
-import { z, schemaForType } from "../../../util/zodUtil";
+import { z, schemaForType } from "../../../parse";
 
 export enum Theme {
   DARK = "dark",
